test deploy

+++ src/main/java/com/nt/subscribtion_data/model/dao/DataModel/EventData/EventItem/Offer.java
@@ -76,7 +76,7 @@ public class Offer {
     private String descriptionEn;
 
     @JsonProperty("ServiceType")
-    private Integer ServiceType;
+    private Long ServiceType;
 
     @JsonProperty("ocsOfferingName")
     private String ocsOfferingName;
